Fix float32 precision loss in trajectory entropy histogram
compute_trajectory_entropy force-cast actions to float32 before
np.histogram, which can collapse distinct values (and crash with
"Too many bins for data range") for large-offset, sub-millimeter-
variance action columns -- exactly the near-duplicate-trajectory
case this function is meant to detect. Reported via Reddit user
testing on real data.

+++ calibra/curation/entropy.py
@@ -65,7 +65,9 @@ def compute_trajectory_entropy(actions: np.ndarray, num_bins: int = 20) -> float
     joint). This is fast and sufficient for detecting near-duplicate trajectories
     where all dimensions are simultaneously low-entropy.
     """
-    acts = np.asarray(actions, dtype=np.float32)
+    # float32 loses precision on the sub-mm-variance columns this function targets,
+    # which can collapse distinct values and make np.histogram raise on degenerate edges.
+    acts = np.asarray(actions, dtype=np.float64)
     if acts.ndim == 1:
         acts = acts[:, np.newaxis]
